+++ OnePlat.DiceNotation/Dice.cs
@@ -48,7 +48,7 @@ namespace OnePlat.DiceNotation
         }
 
         /// <inheritdoc/>
-        IDice IDice.Dice(int sides, int numberDice, int scalar, int? choose)
+        IDice IDice.Dice(int sides, int numberDice, double scalar, int? choose)
         {
             this.terms.Add(new DiceTerm(numberDice, sides, scalar, choose));
             return this;
